encoding done

+++ src/coding/DCT.hpp
@@ -10,16 +10,16 @@ class Block4x4Layout;
 
 //void multiple(double **block, double matrix[][4], double **out, MultipMode mode);
 
-void idct(double **block);
+void idct(int **block);
 
-int dct(double **block);
+int dct(int **block);
 
 void dct(IBlock4x4& block);
 
 void dct(Block4x4_float &block);
 
-void idct_dc(double **block);
+void idct_dc(int **block);
 
-int dct_dc(double **block);
+int dct_dc(int **block);
 
 #endif //DCT_HPP
